Update urls to reflect move from Dataset to DatasetSubmissionController
These are urls I missed in the initial work of moving submission actions into DatasetSubmissionController.

It will also fix a 500 error when connecting to the private dataset url
for a dataset created through the wizard
because the wizard doesn't set the publisher_id property of dataset and the view tries to
show publisher name without checking whether it is set.