Configure docker-postinstall role with GitLab project name
The GitLab project name within the GitLab URL is hard-coded into the
docker-postinstall/tasks/main.yml file. To fix this, the URL for
the GitLab project is now located in the Ansible
ops/infrastructure/group_vars/all/vault file and it is referred to
by the GitLab URL variable (gitlab_url) which is referenced in the
ops/infrastructure/inventories/hosts file.

The gitlab_url variable is now used within the docker-postinstall
role in its tasks/main.yml file instead of the hard-coded GitLab URL.

This commit addresses GitHub issue #282.