Update Create Dataset Author tab
The Create Dataset Author tab has been updated with functionality to allow a delimited file containing author information to be uploaded. CsvHelper has been created for parsing author information in CSV and TSV files. Also, contribution information can also be added for authors. The database schema has been updated with a new CONTRIBUTION table which is linked to the AUTHOR table with a contribution_id column.

Resolves WL Sprint 1 task: https://jirasoft.ffflabel-dev.com/browse/GIG-5 for issue #241