Fix permissions bugs that caused ftp upload and database update to fail
Make sure that the GID of in-container process is the same on ftpd, watcher and app services.
and that created directories are group writable.

All the actions that need to be triggered upon change to the upload directory
are now performed in one single script (process-uploads.php) to make the work flow less brittle.