Detect appropriate docker hostname for linux or mac/windows
on gitlab, the host name to reach docker daemon for container-host communication
is different, so we need to detect that during configuration of FUW, so appropriate
host name is configured for FUW Yii2 app as a parameter.

Also unit tests needs to be able to read Yii::$app params configuration.