Fix dataset redirect link not functioning with long urls
This looks like and old bug rather than a regression. It may happen on production too.

There was a framework level validation limiting max characters to 50 only.
The acceptance tests didn't fail and pick that up because the url used in the scenario
was unfortunately exactly 50 characters.

Fixed the test scenario and the validation to match the database column max character.