Add File Upload Wizard user creation to TokenService for authorisation
TokenService was created to handle authorisation between GigaDB and File Upload Wizard.
So it's a fitting location for creating the user identity required
to authenticate the dataset author  who upload files with File Upload Wizard.

It's a fairly standalone change that's part of the bigger work involved in authenticating such user.
That's why I commit it here (small and standalone and it simplifies the upcoming commit of the bigger piece of work).

TokenService is probably not a good name anymore. Will change it later (TODO), may be to AuthorisationService.