Add a getter function for retrieval of dataset submitter to DatasetDAO
As part of the bigger piece of work of authenticating and authorising
the author of a dataset to upload files using File Upload Wizard, we need
to retrieve the dataset's submitter user id form the server side, to compare
dataset associate with filedrop account and author accessing the uploader
and metadata form.

So we need to add an additional getter to the DatasetDAO to get the associated submitter.
Similarly to the previous commit, this is a fairly small and standalone change that
can be committed as is, to simplify the upcoming bigger commit for the larger aforementioned
piece of work.