Update GigaDB file upload author's view to use Javascript Vue components
The Gigadb javascript application (located in ``gigadb/app/client/web``) for uploading files built and then deployed
in ``js/`` directory is added to the view by registering it as client script
in the Yii view layout (uploader_layout.php) used by the view for AuthorisedDatasetController.php's filesUpload.php view.

In the view, we make sure we have a <div> with the same id (#gigadb-fuw) as the Javascript application
(see ``gigadb/app/client/web/src/index.js``) for the Javascript app to work in the page
and inside we can use the JS component
(see ``gigadb/app/client/web/src/components/DatasetInfoComponent.vue``)
as a Custom Element (<dataset-info>).

Updated acceptance test scenario to reflect the variation in terminology so we can show the different stages
in File Upload Wizard.

Updated the FUW docs to indicate the command to test, build and deploy the Javascript application:
```
$ docker-compose run --rm js
```

The deployed Javascript application (js/fuw-*) is .gitignored as it's a build artefact
and dependencies integrity is performed with ``gigadb/app/client/web/package-lock.json``.

Some work in progress (@wip) for acceptance scenario and its implementation was added
to test the flow between the file uploading stage and the file meta-data annotation stage.

Improved the CSS on ``filesUpload.php`` to match the GigaDB's new look for public pages.