Start removing hardcoded DSN from acceptance tests (from latest change)
The postgresql DSN is hard-coded all other the place in the acceptance test
context classes instead of reading them from config.
Because it's quite tangled
(there are 3 ways these classes are connecting to the database)
we do it in piece started from the latest change in GigaDBWebsiteContext class.

For scenario steps we can use an instance property set in the constructor
as the instance is accessible in steps.
This approach will be used for other parts of the code later.