Update file upload acceptance test scenario for better accuracy
The scenario that was uploading files using File Upload Wizard interface
was not testing success at the backend of the Uppy uploading widget,
only that the widget was uploading the files successfully to the TUS server.

Now we check that the files uploads have landed in the FUW database associated with the
correct DOI.

Also we need to remove the uploads entries created during the scenario run
after it has run.