Add more fixes to the between components event passing mechanism
Two concomitant issues are fixed here.

In UploaderComponent.vue, the event to set the uploading stage was not
emitted because when the component is in mounted lifecycle, it doesn't mean
all the dependencies of such component are ready.
So the event emitting code need to be wrapped in a nextTick block to ensure
the event is emitted only when the component and all its dependencies are
mounted and ready.

The second issues is that the PagerComponent.vue was receving UploaderComponent.vue's events
successfully but the instance variables were not updated accordingly.

This is because the "this" keyword loses its meaning when called in a generic
anonymous function like event handlers.

"this" represent the Vue instanced when use in anonymous functions attached to
Vue lifecycle, computed variables and data block but if within those,
we use another anonymous function (with the 'function' keyword), "this"
is then attached to this new context and not to the Vue instance anymore.

The fix we chose is to use a clojure whereby we save "this" into a variable ('vm')
that would be used in the inner event handler.