Add passing acceptance test for showing "Next" when upload is complete
Using the event passing mechanism between UploaderComponent.vue
and PagerComponent.vue, we can display a Next button when we are in
uploading stage and files have been successfully uploaded.

Also re-organised the in-document template for the main Vue app (which
is also the view file for the PHP controller authorisedDataset's filesUploadAction)
to be more semantically correct and to make use of the custom element provided
by our Vue components.