Create new component for the metadata form of the File Upload Wizard
The AnnotationComponent.vue will use ajax to connect to the File Upload Wizard API
to get the list of files uploaded to the dataset and display them in a table.

Axios, used in the component, is the recommended library for making ajax request.

Moxios, used in the unit test file, is library to help testing code that use Axios.

Before moxios comes into play, we use a Jasmine spy to test that axios's ``get`` method
is called with the correct url to our api.

It's still work in progress.