Enable access to /uploads endpoint of File Upload Wizard from PHP
Following a change in approach to the loading of file uploads into metadata form,
in which we will supply the list of uploads to the form from server side,
we need to enable the uploads endpoint to be accessed in-container from
the PHP code.

That's why we are adding a client getter to FiledropService,
and enable the fuw-public-api internal domain in docker-compose.yml

(similarly to fuw-backend, we will need to enable the fuw-frontend.dev.http.conf when starting the app).