Register AnnotatorComponent and improve php to javascript data passing
I forgot to register our new Vue component for the metadata form (AnnotatorComponent.vue) into the main app (index.js)

Also we need to pass data from server side php to the annotator component.
This is done by passing the PHP array as json to the Annotator component's ``uploads`` props.
the ``uploads`` attribute needs to be prefixed with Vue's 'v-bind' command so to recognize the value as
a Javascript expression.

Added a test to AnnotatorComponent for ensuring the ``uploads`` props is set.