Fix missing data type when loading list of uploads in metadata form
the data type was not saved in database upon the files being uploaded by the user.
(the code was commented out because at that time, the datatype field wasn't added yet)
So fixed the file-to-db.php script that's triggered when the files are
physically written the drop box file system by ensuring the data type
is calculated and inserted in the database record.

Also quick fix the list of mappings between extension and data type to be more realistic.