Create REST endpoint in FileUploadWizard to send email notifications
this is so GigaDB have a REST endpoint (POST /notifications/emailSend) in FUW to connect to to send email
(this will be used to send email when dataset status changes to 'DataAvailableForReview')

The new EmailSendAction.php action from the new NotificationController.php in the ``frontend`` Yii2 app
(in there because it is triggered by non-admin users)
will reuse the existing MessagingService.php component. However that service was in the ``backend`` Yii2 app
as it was created for the admin task of creating filedrop account and notifiying authors.
So we need to move that component (and its corresponding automated test suite) to the ``common`` Yii2 app
which remit is to hold the common code used by all Yii2 app.